Migrate old LSP configs to mason
This significantly cleans up the LSP config, as well as automatically
installs the relevant language servers. Setup in the simple case is
automatic, and there's no need to explicity set up each one.

The non-lspconfig parts (goimports and on_attach) were moved to
golang.lua and lsp_on_attach.lua respectively, and the goimports
function was updated with the latest recommendations from the gopls
website.